backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Vector Database
chromadb>=0.4.0

# Machine Learning
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10

# OpenAI API
openai>=1.3.0

# Configuration Management
python-dotenv>=1.0.0
keyring>=24.0.0

# Frontmatter Parsing
python-frontmatter>=1.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.88.0

# Development
pytest-asyncio>=0.21.0

# Menu Bar Application
pystray>=0.19.0
Pillow>=9.0.0
